+++ CHANGELOG.md
@@ -1,7 +1,19 @@
 ## Upcoming Changes
 ### Breaking
+
+### Features
+* timestamper: support generating timestamps from the current time when no `source_fields` are configured
+
+### Improvements
+* timestamper: validate `source_format` and `source_timezone` according to the configured `source_fields`
+
+### Bugfix
+
+## 21.0.0
+### Breaking
 * restrict UNIX timestamp normalization to seconds, milliseconds, microseconds, and nanoseconds.
 * handle non-string `concatenator` source fields with `ProcessingWarning` instead of `ProcessingCriticalError`
+* filter: field values are now automatically coerced in range matches
 
 ### Features
 * add support for fractional UNIX timestamps in the `timestamper` processor while preserving supported integer timestamp normalization.
@@ -9,6 +21,8 @@
 * generic_adder: add support for templated http urls & content_field
 * generic_resolver: add content_field support
 * field_name_replacer: add new `field_name_replacer` processor to replace occurences of strings in key names
+* filter: add support for `*` (open boundary) in range expressions
+* filter: allow mixed numeric range boundaries and type coercion for range matching
 
 ### Improvements
 * docs: enable pydoc placeholders for facilitating component reuse through inheritance
@@ -24,13 +38,17 @@
 * ci: enforce CHANGELOG.md is updated
 * ci: enforce PR TODOs are completed
 * ci: introduce umbrella job for enforcing merge status checks with GitHub
+* grokker: improve performance by stopping on first matching expression
 
 ### Bugfix
 * chart: fix command handling
 * ng: fix input timeout to also accept int parameters
 * ng: fix `http_input` `collect_meta` leading to shared dicts between events
 * ng: fix error output structure to stay consistent with non-ng
-* generic_adder: allow None as valid input via `add`
+* generic_adder: allow `None` as valid input via `add`
+* grokker: allow fallback matches without named fields
+* filter: fix lucene range expressions not matching on mixed-type scenarios
+* filter: treat `inf`/`nan` as string values instead of numeric range boundaries
 
 ## 20.0.0
 ### Breaking